requirements.txt
argcomplete==1.12.0
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
bottle==0.12.18
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.1.1
djangorestframework==3.12.0
idna==2.10
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.5.0
packaging==20.4
pipx==0.15.5.1
pluggy==0.13.1
prettytable==0.7.2
py==1.9.0
PyHamcrest==2.0.2
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pytest==5.4.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
userpath==1.4.1
wcwidth==0.2.5
wrapt==1.12.1
